// frontend/viewer/src/lib/dotnet-types/generated-types/FwLiteShared/Services/IJsInvokableLogger.ts
/* eslint-disable */
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import type {LogLevel} from '../../Microsoft/Extensions/Logging/LogLevel';

export interface IJsInvokableLogger
{
	log(level: LogLevel, message: string) : Promise<void>;
}
/* eslint-enable */