// frontend/viewer/src/lib/dotnet-types/generated-types/Microsoft/Extensions/Logging/LogLevel.ts
/* eslint-disable */
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export enum LogLevel {
	Trace = 0,
	Debug = 1,
	Information = 2,
	Warning = 3,
	Error = 4,
	Critical = 5,
	None = 6
}
/* eslint-enable */